fix: cross-compiling the shell for Arm architecture
FossilOrigin-Name: eefe548787738dd6ff35288de3850ac181ad887f06a09bb2ceadc5b262e3d80d

+++ src/variables.nim
@@ -130,7 +130,7 @@ proc setVariables*(newDirectory: DirectoryPath; db;
     newDirectory.len > 0
     db != nil
   body:
-    var skipped: seq[int] = @[]
+    var skipped: seq[int64] = @[]
 
     # Remove the old environment variables if needed
     if oldDirectory.len > 0:
